test(examples): clean up & fix "things" and their tests
